requirements.txt
-i https://pypi.org/simple
apipkg==1.5
appium-python-client==1.3.0
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
decorator==5.0.6
docutils==0.17
execnet==1.8.0
idna==2.10
importlib-metadata==3.10.0 ; python_version < '3.8'
iniconfig==1.1.1
kitchen==1.2.6
more-itertools==8.7.0
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest-forked==1.3.0
pytest-xdist==2.2.1
pytest==6.0.0
pyyaml==5.4.1
requests==2.25.1
sauceclient==1.0.0
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3 ; python_version < '3.8'
urllib3==1.26.4
zipp==3.4.1
